changed the border witdh spinner increment to 0.25 instead of 1, like it was in iReport
git-svn-id: http://code.jaspersoft.com/svn/repos/jaspersoftstudio/trunk@7738 6833c83e-4fd0-4c68-a3df-283c6148db30

+++ com.jaspersoft.studio/src/com/jaspersoft/studio/swt/widgets/NullableSpinner.java
@@ -423,7 +423,7 @@ public class NullableSpinner extends Composite {
 	 * @param value the new increment (must be greater than zero)
 	 * 
 	 */
-	public void setIncrement(int value) {
+	public void setIncrement(float value) {
 		this.checkWidget();
 		text.setIncrementStep(value);
 	}
